+++ src/layout/Layout.tsx
@@ -8,6 +8,8 @@ import { Notification } from "../common/Notification";
 import { ViewPortListener } from "../common/ViewPortListener";
 import { assetInfoStore } from "../store/assetInfoStore";
 import { network } from "../constants/constants";
+import { viewPortStore } from "../store/viewPortStore";
+import { notificationStore } from "../store/notificationStore";
 
 type LayoutProps = {
   children: ReactNode;
@@ -15,11 +17,14 @@ type LayoutProps = {
 
 export const Layout: FC<LayoutProps> = ({ children }) => {
   const { fetchAndSetAssetsAndChains } = assetInfoStore();
+  const { isMobile } = viewPortStore();
   const theme = getCurrentTheme();
+  const { fetchNotification } = notificationStore();
 
   useEffect(() => {
     fetchAndSetAssetsAndChains();
-  }, [fetchAndSetAssetsAndChains]);
+    fetchNotification();
+  }, [fetchAndSetAssetsAndChains, fetchNotification]);
 
   return (
     <div className={`${theme} relative overflow-hidden bg-opacity-50`}>
@@ -32,7 +37,7 @@ export const Layout: FC<LayoutProps> = ({ children }) => {
           <Navbar />
           {children}
         </div>
-        <Notification />
+        {!isMobile && <Notification />}
         <Footer className={"mt-auto"} network={network} />
       </div>
     </div>